requirements.txt
Flask
pymongo
Flask-SQLAlchemy
requests
psycopg2-binary
